add menuContent and fix bug
