+++ src/pages/AddPage/AddPage.js
@@ -11,7 +11,8 @@ import {
 } from '../../styles/icon'
 import flexJustifyAlign from '../../styles/flexJustifyAlign'
 import { BackstageNavbar } from '../../components/BackstageNavbar'
-// import Menu from '../../components/Menu'
+import { Menu } from '../../components/Menu'
+import BackstageMenuContent from '../../components/BackstageMenuContent'
 
 const Wrapper = styled.div`
   top: 4rem;
@@ -33,7 +34,7 @@ const CardWrapper = styled.div`
 
 const Card = styled(Link)`
   width: 100%;
-  padding: 2rem 1rem;
+  padding: 1.8rem 1rem;
   ${flexJustifyAlign()}
   color: ${({ theme }) => theme.secondary_000};
   box-shadow: ${({ theme }) => theme.boxShadow};
@@ -42,7 +43,7 @@ const Card = styled(Link)`
   cursor: pointer;
 
   svg {
-    margin: 0 1rem;
+    margin: 0 0.5rem;
   }
 `
 
@@ -65,7 +66,11 @@ const AddPage = () => {
   const theme = useTheme()
   return (
     <Wrapper>
-      {/* <Menu userId={1} nickname="嘎嘎嗚拉拉" /> */}
+      <Menu
+        userId={1}
+        nickname="嘎嘎嗚拉拉"
+        MenuContent={BackstageMenuContent}
+      />
       <BackstageNavbar iconName={plusIcon} title="建立" />
       <CardWrapper>
         <Card to="/form">
@@ -77,21 +82,21 @@ const AddPage = () => {
             </CardText>
           </CardContent>
         </Card>
-        <Card>
+        <Card to="/add">
           <CardContent>
             <CardTitle>測驗</CardTitle>
             <CardText>創建專屬話題，提供即時留言與您輕鬆地進行話題討</CardText>
           </CardContent>
           {testIcon('3x', theme.primary)}
         </Card>
-        <Card>
+        <Card to="/add">
           {lotteryIcon('3x', theme.primary)}
           <CardContent>
             <CardTitle>抽獎</CardTitle>
             <CardText>創建專屬話題，提供即時留言與您輕鬆地進行話題討</CardText>
           </CardContent>
         </Card>
-        <Card>
+        <Card to="/add">
           <CardContent>
             <CardTitle>問卷</CardTitle>
             <CardText>創建專屬話題，提供即時留言與您輕鬆地進行話題討</CardText>
